Move transaction interfaces into transaction package
