+++ java/core/src/main/java/sleeper/core/statestore/transactionlog/TransactionLogSnapshotCreator.java
@@ -18,6 +18,7 @@ package sleeper.core.statestore.transactionlog;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import sleeper.core.statestore.transactionlog.transactions.StateStoreTransaction;
 import sleeper.core.table.TableStatus;
 
 import java.util.Optional;
