+++ java/core/src/main/java/sleeper/core/statestore/transactionlog/transactions/AddFilesTransaction.java
@@ -18,7 +18,6 @@ package sleeper.core.statestore.transactionlog.transactions;
 import sleeper.core.statestore.AllReferencesToAFile;
 import sleeper.core.statestore.StateStoreException;
 import sleeper.core.statestore.exception.FileAlreadyExistsException;
-import sleeper.core.statestore.transactionlog.FileReferenceTransaction;
 import sleeper.core.statestore.transactionlog.state.StateStoreFile;
 import sleeper.core.statestore.transactionlog.state.StateStoreFiles;
 import sleeper.core.table.TableStatus;
