+++ java/core/src/main/java/sleeper/core/statestore/transactionlog/transactions/ClearFilesTransaction.java
@@ -16,7 +16,6 @@
 package sleeper.core.statestore.transactionlog.transactions;
 
 import sleeper.core.statestore.StateStoreException;
-import sleeper.core.statestore.transactionlog.FileReferenceTransaction;
 import sleeper.core.statestore.transactionlog.state.StateStoreFiles;
 
 import java.time.Instant;
