+++ java/core/src/main/java/sleeper/core/statestore/transactionlog/transactions/SplitPartitionTransaction.java
@@ -17,7 +17,6 @@ package sleeper.core.statestore.transactionlog.transactions;
 
 import sleeper.core.partition.Partition;
 import sleeper.core.statestore.StateStoreException;
-import sleeper.core.statestore.transactionlog.PartitionTransaction;
 import sleeper.core.statestore.transactionlog.state.StateStorePartitions;
 
 import java.time.Instant;
